requirements.txt
click==8.1.7
requests==2.32.3
PyYAML==6.0.1
tqdm==4.67.1
